"Refactor list_files tool to separate module (#2057)"
