+++ src/core/tools/listFilesTool.ts
@@ -1,0 +1,78 @@
+import * as path from "path"
+import { Cline } from "../Cline"
+import { ClineSayTool } from "../../shared/ExtensionMessage"
+import { ToolParamName, ToolUse } from "../assistant-message"
+import { formatResponse } from "../prompts/responses"
+import { listFiles } from "../../services/glob/list-files"
+import { getReadablePath } from "../../utils/path"
+import { AskApproval, HandleError, PushToolResult } from "./types"
+/**
+ * Implements the list_files tool.
+ *
+ * @param cline - The instance of Cline that is executing this tool.
+ * @param block - The block of assistant message content that specifies the
+ *   parameters for this tool.
+ * @param askApproval - A function that asks the user for approval to show a
+ *   message.
+ * @param handleError - A function that handles an error that occurred while
+ *   executing this tool.
+ * @param pushToolResult - A function that pushes the result of this tool to the
+ *   conversation.
+ * @param removeClosingTag - A function that removes a closing tag from a string.
+ */
+export async function listFilesTool(
+	cline: Cline,
+	block: ToolUse,
+	askApproval: AskApproval,
+	handleError: HandleError,
+	pushToolResult: PushToolResult,
+	removeClosingTag: (tag: ToolParamName, text?: string) => string,
+) {
+	const relDirPath: string | undefined = block.params.path
+	const recursiveRaw: string | undefined = block.params.recursive
+	const recursive = recursiveRaw?.toLowerCase() === "true"
+	const sharedMessageProps: ClineSayTool = {
+		tool: !recursive ? "listFilesTopLevel" : "listFilesRecursive",
+		path: getReadablePath(cline.cwd, removeClosingTag("path", relDirPath)),
+	}
+	try {
+		if (block.partial) {
+			const partialMessage = JSON.stringify({
+				...sharedMessageProps,
+				content: "",
+			} satisfies ClineSayTool)
+			await cline.ask("tool", partialMessage, block.partial).catch(() => {})
+			return
+		} else {
+			if (!relDirPath) {
+				cline.consecutiveMistakeCount++
+				pushToolResult(await cline.sayAndCreateMissingParamError("list_files", "path"))
+				return
+			}
+			cline.consecutiveMistakeCount = 0
+			const absolutePath = path.resolve(cline.cwd, relDirPath)
+			const [files, didHitLimit] = await listFiles(absolutePath, recursive, 200)
+			const { showRooIgnoredFiles = true } = (await cline.providerRef.deref()?.getState()) ?? {}
+			const result = formatResponse.formatFilesList(
+				absolutePath,
+				files,
+				didHitLimit,
+				cline.rooIgnoreController,
+				showRooIgnoredFiles,
+			)
+			const completeMessage = JSON.stringify({
+				...sharedMessageProps,
+				content: result,
+			} satisfies ClineSayTool)
+			const didApprove = await askApproval("tool", completeMessage)
+			if (!didApprove) {
+				return
+			}
+			pushToolResult(result)
+			return
+		}
+	} catch (error) {
+		await handleError("listing files", error)
+		return
+	}
+}
